envoy.distribution.release: Rename publish -> release
