--- envoy.distribution.publish/README.rst
@@ -1,5 +1,0 @@
-
-envoy.distribution.publish
-==========================
-
-Package publishing tool used in Envoy proxy's CI
