+++ envoy.github.release/VERSION
@@ -1,1 +1,1 @@
-0.0.3
+0.0.4-dev
